sum_api/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.3
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.3.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
miniful==0.0.6
numpy==1.24.2
pipdeptree==2.5.0
pyFUME==0.2.25
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
simpful==2.10.0
six==1.16.0
smart-open==6.3.0
threadpoolctl==3.1.0
urllib3==1.26.14
Werkzeug==2.2.3
azure-functions
